+++ src/styled_char.rs
@@ -22,7 +22,7 @@ use {
 
 /// A modifiable character which can be easily written or repeated. Can
 /// be used for bullets, horizontal rules or quote marks.
-#[derive(Clone, Debug, PartialEq)]
+#[derive(Clone, Copy, Debug, PartialEq)]
 pub struct StyledChar {
     compound_style: CompoundStyle,
     nude_char: char,